+++ airbnb/airbnb/Sources/Model/Lodging.swift
@@ -1,0 +1,14 @@
+//
+//  Lodging.swift
+//  airbnb
+//
+//  Created by seongha shin on 2022/05/25.
+//
+
+import Foundation
+
+struct Lodging: Decodable {
+    let name: String
+    let coordX: Double
+    let coordY: Double
+}
